refactor: move EventBus to dedicated namespace
