+++ CodexTest/Assets/Scripts/Infrastructure/ClientInputSender.cs
@@ -4,6 +4,7 @@ using Game.Domain.Events;
 using Game.Networking;
 using Game.Networking.Messages;
 using Game.Utils;
+using Game.EventBus;
 using UnityEngine;
 using UnityEngine.InputSystem;
