+++ CodexTest/Assets/Scripts/Infrastructure/ServerBootstrap.cs
@@ -8,6 +8,7 @@ using Game.Systems;
 using Game.Utils;
 using System.Collections.Generic;
 using Unity.Networking.Transport;
+using Game.EventBus;
 using UnityEngine;
 
 namespace Game.Infrastructure
