+++ CodexTest/Assets/Scripts/Systems/SurvivalSystem.cs
@@ -1,7 +1,7 @@
 using Game.Components;
 using Game.Domain.ECS;
 using Game.Domain.Events;
-using Game.Infrastructure;
+using Game.EventBus;
 using System.Linq;
 using UnityEngine;
