+++ CodexTest/Assets/Tests/CommandHandlerTests.cs
@@ -1,7 +1,7 @@
 using Game.Domain;
 using Game.Domain.Commands;
 using Game.Domain.ECS;
-using Game.Infrastructure;
+using Game.EventBus;
 using NUnit.Framework;
 using UnityEngine;
