+++ CodexTest/Assets/Tests/JumpSystemTests.cs
@@ -3,7 +3,7 @@ using Game.Components;
 using Game.Domain.Commands;
 using Game.Domain.ECS;
 using Game.Domain.Events;
-using Game.Infrastructure;
+using Game.EventBus;
 using Game.Systems;
 using NUnit.Framework;
 using UnityEngine;
