+++ CodexTest/Assets/Tests/ReplicationSystemTests.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using Game.Domain.ECS;
 using Game.Domain.Events;
-using Game.Infrastructure;
+using Game.EventBus;
 using Game.Networking;
 using Game.Networking.Messages;
 using Game.Systems;
