+++ CodexTest/Assets/Tests/SurvivalSystemTests.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using Game.Components;
 using Game.Domain.ECS;
 using Game.Domain.Events;
-using Game.Infrastructure;
+using Game.EventBus;
 using Game.Systems;
 using NUnit.Framework;
